PersonList: split filter into its own component
